Fix stack overflow error in RegexAutomaton's emptyLanguage, abstract semantics of repeat for Tarsis, testing single string repetition

+++ lisa/lisa-sdk/src/main/java/it/unive/lisa/util/datastructures/regex/Atom.java
@@ -184,12 +184,12 @@ public class Atom extends RegularExpression {
 	}
 
 	@Override
-	public RegularExpression repeat(int n) {
+	public RegularExpression repeat(long n) {
 		if (n == 0)
 			return Atom.EPSILON;
 
 		RegularExpression r = Atom.EPSILON;
-		for (int i = 0; i < n; i++)
+		for (long i = 0; i < n; i++)
 			r = new Comp(r, this);
 		r.simplify();
 		return r;
