+++ lisa/lisa-sdk/src/main/java/it/unive/lisa/util/datastructures/regex/EmptySet.java
@@ -134,7 +134,7 @@ public final class EmptySet extends RegularExpression {
 	}
 
 	@Override
-	public RegularExpression repeat(int n) {
+	public RegularExpression repeat(long n) {
 		return this;
 	}
 }
